# Optimize abstract skimmer