--- a/src/main/scala/beam/agentsim/infrastructure/ChargingNetworkManager.scala
+++ b/src/main/scala/beam/agentsim/infrastructure/ChargingNetworkManager.scala
@@ -64,7 +64,6 @@ class ChargingNetworkManager(
   private def currentTimeBin(tick: Int): Int = cnmConfig.timeStepInSeconds * (tick / cnmConfig.timeStepInSeconds).toInt
   private def nextTimeBin(tick: Int): Int = currentTimeBin(tick) + cnmConfig.timeStepInSeconds
 
-
   var timeSpentToPlanEnergyDispatchTrigger: Long = 0
   var nHandledPlanEnergyDispatchTrigger: Int = 0
 
@@ -144,7 +143,7 @@ class ChargingNetworkManager(
       val e = System.currentTimeMillis()
       nHandledPlanEnergyDispatchTrigger += 1
       timeSpentToPlanEnergyDispatchTrigger += e - s
-      
+
       sender ! CompletionNotice(triggerId, triggers.toIndexedSeq ++ nextStepPlanningTriggers)
 
     case TriggerWithId(ChargingTimeOutTrigger(tick, vehicleId, vehicleManager), triggerId) =>