--- a/src/main/scala/beam/router/skim/core/DriveTimeSkimmer.scala
+++ b/src/main/scala/beam/router/skim/core/DriveTimeSkimmer.scala
@@ -57,12 +57,12 @@ class DriveTimeSkimmer @Inject()(
                 val key = PathCache(origin.tazId, destination.tazId, timeBin)
                 observedTravelTimes.get(key).foreach { timeObserved =>
                   val theSkimKey = DriveTimeSkimmerKey(origin.tazId, destination.tazId, timeBin * 3600)
-                  currentSkimInternal.get(theSkimKey).map(_.asInstanceOf[DriveTimeSkimmerInternal]).foreach {
+                  getCurrentSkimValue(theSkimKey).map(_.asInstanceOf[DriveTimeSkimmerInternal]).foreach {
                     theSkimInternal =>
                       series += ((theSkimInternal.observations, theSkimInternal.timeSimulated, timeObserved))
                       for (_ <- 1 to theSkimInternal.observations)
                         deltasOfObservedSimulatedTimes += theSkimInternal.timeSimulated - timeObserved
-                      currentSkimInternal.update(theSkimKey, theSkimInternal.copy(timeObserved = timeObserved))
+                      currentSkimInternal.put(theSkimKey, theSkimInternal.copy(timeObserved = timeObserved))
                   }
                 }
               }