Latest production run on smart-baseline scenario with twice lower population has bad performance for 0-th iteration. Decided to add some logs  and collect stats from profiler. Here what I saw:
![image](https://user-images.githubusercontent.com/5107562/109707457-c4c93e80-7bcc-11eb-941d-5c855ae2e915.png)

```
15:54:50.193 [ClusterSystem-akka.actor.default-dispatcher-90] INFO  b.a.scheduler.BeamAgentScheduler - Hour 59.5 completed. 73.3(GB)
15:55:02.950 [ClusterSystem-akka.actor.default-dispatcher-100] INFO  b.a.i.ChargingNetworkManager - timeSpentToPlanEnergyDispatchTrigger: 2624753 ms, nHandledPlanEnergyDispatchTrigger: 717, AVG: 3660.7433751743374
15:55:15.516 [ClusterSystem-akka.actor.default-dispatcher-65] INFO  b.a.scheduler.BeamAgentScheduler - Hour 60.0 completed. 91.3(GB)
15:55:15.517 [ClusterSystem-akka.actor.default-dispatcher-17] INFO  b.a.scheduler.BeamAgentScheduler - Stopping BeamAgentScheduler @ tick 216001. Iteration 0 executed in 3309 seconds
```
0-th iteration took 3309 seconds, but ChargingNetworkManager on it's own it spent 2624 seconds handling PlanEnergyDispatchTrigger message.
If I look to the profiler, I can see it spend a lot of time creating [copy of mutable hasmap](https://github.com/LBNL-UCB-STI/beam/blob/f570ce5c95e246e63e8d00d66f91dfa40c966c56/src/main/scala/beam/router/skim/core/AbstractSkimmer.scala#L46):
![image](https://user-images.githubusercontent.com/5107562/109707397-b4b15f00-7bcc-11eb-8e8b-2141d2dbef16.png)

Link to the run results with logs: https://s3.us-east-2.amazonaws.com/beam-outputs/index.html#output/sfbay/sfbay-smartbase__2021-03-02_14-59-13_gpk/

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/lbnl-ucb-sti/beam/3131)
<!-- Reviewable:end -->