@REASY 2021-03-02
CPU usage before:
![image](https://user-images.githubusercontent.com/5107562/109707773-238eb800-7bcd-11eb-87b3-38a52a60112f.png)

CPU usage after:
![image](https://user-images.githubusercontent.com/5107562/109707791-29849900-7bcd-11eb-8164-0259e08d94c6.png)